Update devcontainer to Ubuntu 21.04
Also to the C++ container, which includes other useful C++ tools, even if it also includes gcc.

This has a lot of lines changed, since it regenerates from the template, but the main differences are switching to the C++ (instead of base Ubuntu) container, and change to Ubuntu 21.04, which supports Apple Silicon.